changed file structure and updated backend
