+++ backend/.gitignore
@@ -4,6 +4,6 @@ node_modules
 .env.development.local
 .env.test.local
 .env.production.local
-
+package-lock.json
 
 .env
