--- backend/middleware/authenticateUser.js
@@ -1,28 +1,0 @@
-// // Import the UserModel from the User model file
-// import { UserModel } from "../models/UserModel";
-// // Define a function called authenticateUser that takes a request (req), response (res), and a next function as parameters
-// export const authenticateUser = async (req, res, next) => {
-//   // Retrieve the access token from the request header
-//   const accessToken = req.header("Authorization");
-//   try {
-//     // Find a user in the database using the retrieved access token
-//     // Mongoose Method: UserModel.findOne({ accessToken: accessToken })
-//     // Description: This line of code serves the purpose of authenticating a user based on the provided access token. It checks if a user with the specified accessToken exists in the database using the UserModel. If a user is found, their user document is stored in the user variable. This allows the middleware to add the user object to the request, making it available for subsequent middleware or routes. If no user is found, it prepares to send a 401 Unauthorized response to indicate that the user needs to log in. This code is an essential part of user authentication in the Node.js application and helps control access to protected routes or endpoints.
-//     const user = await UserModel.findOne({ accessToken: accessToken });
-//     if (user) {
-//       // If a user is found, add the user object to the request object
-//       req.user = user; // Add user to the request object
-//       next(); // Continue to the next middleware or route
-//     } else {
-//       // If no user is found, send a 401 Unauthorized response
-//       res.status(401).json({ success: false, response: "Please log in" });
-//     }
-//   } catch (e) {
-//     // Handle any errors that occur during the database query or user authentication
-//     res.status(500).json({ success: false, response: e.message });
-//   }
-// };
-
-// // SUMMARY
-
-// //In this code, we have a function called authenticateUser that is used as middleware in a Node.js application. This middleware is responsible for checking the authorization header of an incoming request, searching for a user with the provided access token in the database using the UserModel, and adding the user object to the request if found. If no user is found or if there are any errors during the process, appropriate responses are sent back to the client. In summary, this code is handling user authentication by checking the access token in the request header and verifying it against the database to grant access to protected routes or endpoints.
