+++ frontend/src/pages/MyFavorites/MyFavorites.jsx
@@ -1,10 +1,10 @@
 //MY PAGE THAT SHOULD BE DISPLAYING THE LOGGED IN USER'S FAVORITE PLAYGROUNDS
 
 
-import { LogOut } from "../components/LogOut"
-import "./home.css"
+import { LogOut } from "../../components/LogOut" 
 
-export const SecretPage = () => {
+
+export const MyFavorites = () => {
     return (
         <div className="home-container">
     <h1>Welcome to the secret page</h1>
