--- frontend/src/pages/home.css
@@ -1,9 +1,0 @@
-.home-container {
-    display: flex;
-        justify-content: center;
-        align-items: center;
-        flex-direction: column;
-        height: 100vh;
-}
-
-
