+++ frontend/src/routes/routes.jsx
@@ -1,12 +1,12 @@
 // Import the 'Route' component from the 'react-router-dom' library.
 import { Route } from "react-router-dom";
 // Import various page components used as route elements.
-import { Home } from "../pages/Home";
-import { Register } from "../pages/Register";
-import { Login } from "../pages/Login";
-import { ErrorPage } from "../pages/ErrorPage";
-import { SecretPage } from "../pages/SecretPage";
-import  PlaygroundDetails  from "../pages/PlaygroundDetails";
+import { Home } from "../pages/Home/Home";
+import { Register } from "../pages/Register/Register";
+import { Login } from "../pages/Login/Login";
+import { ErrorPage } from "../pages/Errorpage/ErrorPage";
+import  PlaygroundDetails  from "../pages/Playgrounddetails/PlaygroundDetails";
+import { MyFavorites } from "../pages/MyFavorites/MyFavorites";
 
 // Define the 'routes' variable as a JSX expression.
 const routes = (
@@ -18,7 +18,7 @@ const routes = (
     {/* Define a route for the '/home' path that renders the 'Home' component. */}
     <Route path="/login" element={<Login />} />
     {/* Define a route for the '/secretpage' path that renders the 'SecretPage' component, aka the user's favorited playgrounds. */}
-    <Route path="/secretpage" element={<SecretPage />} />
+    <Route path="/secretpage" element={<MyFavorites />} />
     {/* Define a route for the '/register' path that renders the 'Register' component. */}
     <Route path="/register" element={<Register />} />
     {/* Define a catch-all route that matches any other path ('*') and renders the 'NotFound' component. */}
